third_party/blink/web_tests/platform/generic/fast/multicol/vertical-rl/balancing/balance-short-trailing-empty-block-expected.txt

Both words below should be in the the right column.

Tests if a block crossing a column boundary is recorded properly for balancing

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS mc.offsetWidth is 190
PASS successfullyParsed is true

TEST COMPLETE

second
column
